refactor(api/tracker): move errorIdentityFactory file
